Cleanup configuration
