+++ src/main/java/com/activeviam/apps/cfg/datastore/DatastoreSchemaConfig.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) ActiveViam 2024
+ * Copyright (C) ActiveViam 2024-2025
  * ALL RIGHTS RESERVED. This material is the CONFIDENTIAL and PROPRIETARY
  * property of ActiveViam Limited. Any unauthorized use,
  * reproduction or transfer of this material is strictly prohibited
@@ -9,7 +9,6 @@ package com.activeviam.apps.cfg.datastore;
 import java.util.List;
 
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 
 import com.activeviam.activepivot.server.spring.api.config.IDatastoreSchemaDescriptionConfig;
 import com.activeviam.database.datastore.api.description.IDatastoreSchemaDescription;
@@ -19,7 +18,6 @@ import com.activeviam.database.datastore.api.description.impl.DatastoreSchemaDes
 
 import lombok.RequiredArgsConstructor;
 
-@Configuration
 @RequiredArgsConstructor
 public class DatastoreSchemaConfig implements IDatastoreSchemaDescriptionConfig {
     private final List<? extends IStoreDescription> storeDescriptions;
