+++ src/test/java/com/activeviam/apps/cfg/pivot/CubeTestConfig.java
@@ -1,14 +1,14 @@
 /*
- * Copyright (C) ActiveViam 2024
+ * Copyright (C) ActiveViam 2024-2025
  * ALL RIGHTS RESERVED. This material is the CONFIDENTIAL and PROPRIETARY
  * property of ActiveViam Limited. Any unauthorized use,
  * reproduction or transfer of this material is strictly prohibited
  */
 package com.activeviam.apps.cfg.pivot;
 
+import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 
 import com.activeviam.activepivot.core.datastore.api.builder.StartBuilding;
 import com.activeviam.activepivot.core.intf.api.description.IActivePivotManagerDescription;
@@ -17,8 +17,8 @@ import com.activeviam.atoti.server.test.api.CubeTester;
 import com.activeviam.database.datastore.api.IDatastore;
 import com.activeviam.database.datastore.api.description.IDatastoreSchemaDescription;
 
-@Configuration
-@ComponentScan(basePackageClasses = {DatastoreConfig.class, ActivePivotConfig.class})
+@TestConfiguration
+@Import({DatastoreConfig.class, ActivePivotManagerConfig.class, CubeConfig.class})
 public class CubeTestConfig {
     private final CubeTester cubeTester;
     private final IDatastore datastore;
